Server/requirements.txt
fastapi
uvicorn
python-decouple
supabase
python-multipart